Add files via upload
The Dairy Farm Shop Management System (DFSMS) is a web based application that can be
accessed over the web. This system can be used to automate the workflow of dairy shop and
their invoices.
The project has been planned to be having the view of distributed architecture, with centralized
storage of the database. The application for the storage of the data has been planned. Using
the constructs of MySQL Server and all the user interfaces has been designed using the PHP
technologies. The database connectivity is planned using the ”MySQL Connection” method-
ology. The standards of security and data protective mechanism have been given a big choice
for proper usage. The application takes care of different modules and their associated reports.
which are produced as per the applicable strategies and standards that are put forwarded by
the administrative staff.
DFSMS is a web-based application which manages the products of dairy shop. It has one
module i.e. admin who manages all the functions of the dairy shop.